Remove __VA_OPT__ dependency in namespace config
Depending on __VA_OPT__ for namespace configuration is limiting
while it's not supported in all language modes without warnings.

Keeping the single "namespace.hpp" file; considered separating into
"namespace_open.hpp" and "namespace_close.hpp" or changing the name.
This is simplest and can be removed with a pcpp preprocessor.