Deduce index type in subscript ops
subscript(a,i) and flat_index(a,i) now deduce type of i
and default to int=0 for subscript(i) and flat_index(i).
flat_index_recurse(a,i) now deduces type of i
and its default argument i=0 is removed.

With size_t as index type,  warnings of implicit conversion
are given when called with a signed integer.