flat_index unroll to 4D
MSVC struggles with constant evaluation, e.g. in tupl cat.
This seems to be related to 'returns a reference to temporary'
warnings in nested function calls that pass a ref to temp.

Try unrolling multidim array access up to 4D to avoid the nested
recursive helper call.

Also, incorporate the non-array case and remove the overload.

Add documentation about lack of bounds-check.